feat: implement glob pattern path mapping to master
I have implemented a mechanism to map paths with dynamic glob patterns correctly to their corresponding file on the master branch. This change should not be a breaking change, and 0 configuration changes are necessary to implement it. It will just work if you are already using dynamic glob patterns today.

Resolves https://github.com/siddharthkp/bundlesize/issues/208