fix(store): better display for long path names
We have a few quite long path names, for which a flat display doesnt make much sense, so I chose to go for a "vertical" display in all cases

Quick question: when was the last time it was deployed? I still don't see my change from #193 when I go to the online version.

Thanks!